# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "EVA API documentation",
        "title": "Swagger EVA API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "vintobolt",
            "url": "vintobolt",
            "email": "vintobolt@protonmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "host": "127.0.0.1:8089",
    "basePath": "/api/v1/",
    "paths": {
        "/users/signup": {
            "post": {
                "description": "Create a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a user",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Get a user item",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get a user",
                "parameters": [
                    {
                        "enum": [
                            "json",
                            "xml"
                        ],
                        "type": "string",
                        "description": "mediaType",
                        "name": "mediaType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.User": {
            "type": "object",
            "properties": {
                "fullname": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "passwd": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        }
    }
}